% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dongle.R
\name{licd_serial}
\alias{licd_serial}
\title{The dongle's serial}
\usage{
licd_serial(dongle)
}
\arguments{
\item{dongle}{A \code{licd_dongle}.}
}
\value{
The serial as a hex string.
}
\description{
The dongle's serial
}
