% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{licd_strerror}
\alias{licd_strerror}
\title{Text for a status code}
\usage{
licd_strerror(status)
}
\arguments{
\item{status}{One or more status codes, see \link{licd_status_codes}.}
}
\value{
A character vector with the native library's short text for each
code.
}
\description{
Text for a status code
}
\examples{
\dontrun{
licd_strerror(licd_status_codes[["no_device"]])
}
}
